Refactor pixelValue function
Refactor pixelValue function to allow for either one or both zigzag arrays to be used in the code as needed, on the fly. This helps prepare for the future when more patterns will be used, and makes it easier to test patterns.

Make updates to code comments and readme files.